Replace user join/leave opts with registered chans
Implemented RegisterUserJoined & RegisterUserLeft, which take a channel.
This channel is written to, in a blocking fashion, when a user joins or
leaves, to help the consumer reduce potential race conditions. This also
greatly simplified the tests, which were rewritten to improve coverage
and reduce potential blocking. If the channels are nil, they will simply
be ignored.